epic: consolidate commons.js into lib.js
lib.js is the repository convention for shared connector utilities
(commons.js was picked up from legacy connectors like hubspot/xero).
