+++ src/appmixer/epic/bundle.json
@@ -9,7 +9,8 @@
             "Make Api Call component for arbitrary authorized FHIR requests."
         ],
         "1.0.1": [
-            "Production readiness: OAuth2 authorize/token endpoints are now configurable per organisation. Set epicOauthBaseUrl explicitly, or let it derive from epicFhirBaseUrl (standard Epic interconnect layout); the public sandbox remains the default. The OAuth audience (aud) now follows the configured FHIR base URL."
+            "Production readiness: OAuth2 authorize/token endpoints are now configurable per organisation. Set epicOauthBaseUrl explicitly, or let it derive from epicFhirBaseUrl (standard Epic interconnect layout); the public sandbox remains the default. The OAuth audience (aud) now follows the configured FHIR base URL.",
+            "Internal: shared utilities consolidated into lib.js (commons.js removed) to follow the repository convention."
         ]
     }
 }
