--- src/appmixer/epic/commons.js
@@ -1,219 +1,0 @@
-'use strict';
-
-const lib = require('./lib');
-
-// Epic on FHIR R4. The public Epic sandbox base URL is used by default. A
-// different organisation endpoint can be supplied via the instance config
-// (`config.epicFhirBaseUrl`) without changing component code.
-const FHIR_BASE_URL = 'https://fhir.epic.com/interconnect-fhir-oauth/api/FHIR/R4';
-
-// OAuth2 (SMART on FHIR) endpoints for the Epic sandbox.
-const OAUTH_BASE_URL = 'https://fhir.epic.com/interconnect-fhir-oauth/oauth2';
-
-function getFhirBaseUrl(context) {
-
-    const fromConfig = context.config && context.config.epicFhirBaseUrl;
-    return (fromConfig || FHIR_BASE_URL).replace(/\/+$/, '');
-}
-
-// OAuth2 endpoints for the organisation the connector talks to. Every Epic
-// customer hosts its own authorize/token endpoints next to its FHIR API.
-// Precedence: explicit `config.epicOauthBaseUrl` → derived from
-// `config.epicFhirBaseUrl` (Epic's standard interconnect layout puts OAuth at
-// `<interconnect>/oauth2` next to `<interconnect>/api/FHIR/R4`) → the public
-// sandbox default.
-function getOauthBaseUrl(context) {
-
-    const fromConfig = context.config && context.config.epicOauthBaseUrl;
-    if (fromConfig) {
-        return fromConfig.replace(/\/+$/, '');
-    }
-
-    const fhirBase = getFhirBaseUrl(context);
-    if (fhirBase !== FHIR_BASE_URL) {
-        const interconnect = fhirBase.replace(/\/api\/FHIR\/[^/]+$/i, '');
-        if (interconnect !== fhirBase) {
-            return `${interconnect}/oauth2`;
-        }
-    }
-
-    return OAUTH_BASE_URL;
-}
-
-// Perform a FHIR request against the configured base URL using the OAuth2
-// access token. `resource` is the path after the base URL, e.g.
-// `Patient/123` or `Condition`.
-async function fhirRequest(context, { method = 'GET', resource, params } = {}) {
-
-    const response = await context.httpRequest({
-        method,
-        url: `${getFhirBaseUrl(context)}/${resource}`,
-        headers: {
-            'Authorization': `Bearer ${context.auth.accessToken}`,
-            'Accept': 'application/fhir+json'
-        },
-        params
-    });
-
-    return response.data;
-}
-
-// Extract the resources from a FHIR searchset Bundle, skipping OperationOutcome
-// and other non-resource entries.
-function extractResources(bundle) {
-
-    if (!bundle || !Array.isArray(bundle.entry)) {
-        return [];
-    }
-
-    return bundle.entry
-        .map(entry => entry && entry.resource)
-        .filter(resource => resource && resource.resourceType && resource.resourceType !== 'OperationOutcome');
-}
-
-// Shared behaviour for the patient-anchored Find components. Every clinical
-// resource in this connector is searched by patient and returned as a set, so
-// the logic only differs by resource type, the inspector label, the item
-// schema (for the variable picker) and any fixed extra search params (e.g. the
-// Observation category).
-async function runPatientSearch(context, { resourceType, label, schema, extraParams = {} }) {
-
-    const { patient, outputType } = context.messages.in.content;
-
-    if (context.properties.generateOutputPortOptions) {
-        return lib.getOutputPortOptions(context, outputType, schema, { label });
-    }
-
-    if (!patient) {
-        throw new context.CancelError('Patient ID is required!');
-    }
-
-    const bundle = await fhirRequest(context, {
-        resource: resourceType,
-        params: { patient, ...extraParams }
-    });
-
-    const records = extractResources(bundle);
-
-    if (records.length === 0) {
-        return context.sendJson({}, 'notFound');
-    }
-
-    return lib.sendArrayOutput({ context, records, outputType });
-}
-
-// Flatten a FHIR Appointment resource into designer-friendly fields. The full
-// resource is kept under `resource` so no data is lost.
-function normalizeAppointment(resource) {
-
-    const participants = (resource.participant || []).map(p => ({
-        reference: p.actor && p.actor.reference,
-        display: p.actor && p.actor.display,
-        status: p.status
-    }));
-    const byType = prefix => participants.filter(p => (p.reference || '').startsWith(prefix));
-    const patient = byType('Patient/')[0] || {};
-
-    return {
-        id: resource.id,
-        status: resource.status,
-        start: resource.start,
-        end: resource.end,
-        minutesDuration: resource.minutesDuration,
-        created: resource.created,
-        description: resource.description,
-        comment: resource.comment,
-        serviceType: (((resource.serviceType || [])[0] || {}).coding || [{}])[0].display
-            || ((resource.serviceType || [])[0] || {}).text,
-        appointmentType: (resource.appointmentType || {}).text,
-        reason: (((resource.reasonCode || [])[0] || {}).coding || [{}])[0].display
-            || ((resource.reasonCode || [])[0] || {}).text,
-        patient: patient.display,
-        patientReference: patient.reference,
-        practitioners: byType('Practitioner/').map(p => p.display).filter(Boolean),
-        locations: byType('Location/').map(p => p.display).filter(Boolean),
-        participants,
-        lastUpdated: (resource.meta || {}).lastUpdated,
-        resource
-    };
-}
-
-// How far back the polling window reaches. Wide enough to catch updates and
-// cancellations of recent appointments without pulling the whole history.
-const POLL_LOOKBACK_DAYS = 30;
-
-// Shared tick() implementation for the appointment polling triggers.
-// `mode` is one of 'new' | 'updated' | 'cancelled'. The first tick only primes
-// the state (nothing is emitted), subsequent ticks diff against it.
-async function pollAppointments(context, mode) {
-
-    const { patient } = context.properties;
-    if (!patient) {
-        throw new context.CancelError('Patient ID is required!');
-    }
-
-    const since = new Date(Date.now() - POLL_LOOKBACK_DAYS * 24 * 3600 * 1000)
-        .toISOString().slice(0, 10);
-    const bundle = await fhirRequest(context, {
-        resource: 'Appointment',
-        params: { patient, date: `ge${since}` }
-    });
-    const items = extractResources(bundle);
-
-    const state = context.state || {};
-    const known = state.known;
-    const next = {};
-    const changed = [];
-
-    for (const r of items) {
-        const hash = JSON.stringify([r.status, r.start, r.end, r.minutesDuration, (r.meta || {}).lastUpdated]);
-        next[r.id] = { status: r.status, hash };
-        if (!known) continue;
-        const prev = known[r.id];
-        if (mode === 'new' && !prev) {
-            changed.push(r);
-        } else if (mode === 'updated' && prev && prev.hash !== hash) {
-            changed.push(r);
-        } else if (mode === 'cancelled' && r.status === 'cancelled' && (!prev || prev.status !== 'cancelled')) {
-            changed.push(r);
-        }
-    }
-
-    await context.saveState({ known: next });
-
-    for (const r of changed) {
-        await context.sendJson(normalizeAppointment(r), 'out');
-    }
-}
-
-// Flow Test Mode helper for the appointment triggers: return any existing
-// appointment of the patient (normalized) or null when there is none.
-async function findTestAppointment(context) {
-
-    const { patient } = context.properties;
-    if (!patient) {
-        throw new context.CancelError('Patient ID is required!');
-    }
-
-    const since = new Date(Date.now() - POLL_LOOKBACK_DAYS * 24 * 3600 * 1000)
-        .toISOString().slice(0, 10);
-    const bundle = await fhirRequest(context, {
-        resource: 'Appointment',
-        params: { patient, date: `ge${since}` }
-    });
-    const [appointment] = extractResources(bundle);
-    return appointment ? normalizeAppointment(appointment) : null;
-}
-
-module.exports = {
-    FHIR_BASE_URL,
-    OAUTH_BASE_URL,
-    getFhirBaseUrl,
-    getOauthBaseUrl,
-    fhirRequest,
-    extractResources,
-    runPatientSearch,
-    normalizeAppointment,
-    pollAppointments,
-    findTestAppointment
-};
