+++ src/appmixer/epic/core/CancelledAppointment/CancelledAppointment.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { pollAppointments, findTestAppointment } = require('../../commons');
+const { pollAppointments, findTestAppointment } = require('../../lib');
 
 module.exports = {
